Implementacion de rotacion, avance y retroceso

+++ RDX_CONTROL_FIRMWARE_H7/Core/Inc/sbus.h
@@ -13,10 +13,10 @@ void sbus_start(sbus_Handle Handle);
 void sbusParse(sbus_Handle Handle);
 
 
-uint16_t getAcc(sbus_Handle Handle);
+uint32_t getAcc(sbus_Handle Handle);
 uint8_t getDir(sbus_Handle Handle);
 uint8_t sbus_GetState(sbus_Handle Handle);
-uint16_t getGiro(sbus_Handle Handle);
+uint32_t getGiro(sbus_Handle Handle);
 uint16_t getChannel(sbus_Handle Handle , uint8_t num_Channel);
 uint8_t *getBuffer(sbus_Handle Handle);
